Don't parse command line options for skipped files.
